feat: split out agent executable
This allows for a smaller agent executable
that uses less (locked) memory.

Refactor the package `pago` into a number of packages
to enable separating the agent.

Stop locking memory in the main pago executable
when contacting the agent.
It potentially confuses the user's security expectations
because operations without an agent don't do it.

v0.10.0