build(release): use Go arch, OS name for artifacts
There is a case for platform-native, for consistent,
and for standard names.
I chose platform-native names
because they seem more user-friendly in a download list:
quicker and easier for the user of the platform to identify.

There should probably be a layer of indirection
between those names and filenames.

We'll switch to filenames that are consistent
(the same CPU arch is called the same thing everywhere)
and standard for Go.
They can be guessed without going to the download page
and remove the question of naming when we add a new platform.